Combine Bluesky and Mastodon syndication into a single resilient workflow
The separate workflows had three compounding problems:
- npm install modified package.json/package-lock.json, leaving unstaged
  changes that blocked git pull --rebase after the commit
- Because the push never succeeded, bluesky-posted.json was never updated
  on the remote, so every subsequent run reposted the same content
- Two workflows racing on the same trigger could conflict on push

This combines them into a single workflow and script that:
- Runs both platforms in one job (no race condition)
- Saves the posted tracker after each individual post (not at the end),
  so progress is preserved even if the process crashes mid-run
- Uses --no-save for npm install to avoid modifying package.json
- Cleans unstaged changes before rebase, with push retry logic

Also updates bluesky-posted.json to include the 3 posts that were
successfully sent to Bluesky but whose tracker updates never landed.
