+++ .github/data/bluesky-posted.json
@@ -2,6 +2,9 @@
   "posted": [
     "https://www.joehart.co.uk/notes/2026-02-14-45d445a3db93/",
     "https://www.joehart.co.uk/notes/2026-02-14-59846b2d6e85/",
-    "https://www.joehart.co.uk/notes/2026-02-14-9d23a8671e1f/"
+    "https://www.joehart.co.uk/notes/2026-02-14-9d23a8671e1f/",
+    "https://www.joehart.co.uk/notes/2026-02-15-e27c03df6653/",
+    "https://www.joehart.co.uk/notes/2026-02-19-8b0efcd65d30/",
+    "https://www.joehart.co.uk/notes/2026-02-27-eb01e2d215fd/"
   ]
 }
