--- .github/workflows/syndicate-mastodon.yaml
@@ -1,39 +1,0 @@
-name: Syndicate to Mastodon
-
-on:
-  workflow_run:
-    workflows: ["Deploy Now: Orchestration"]
-    types: [completed]
-    branches: [master]
-
-jobs:
-  syndicate:
-    runs-on: ubuntu-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' }}
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Node
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-
-      - name: Wait for deployment propagation
-        run: sleep 60
-
-      - name: Post to Mastodon
-        env:
-          MASTODON_ACCESS_TOKEN: ${{ secrets.MASTODON_ACCESS_TOKEN }}
-        run: node .github/scripts/syndicate-mastodon.js
-
-      - name: Commit posted tracker and backlinks
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add .github/data/mastodon-posted.json notes/
-          if ! git diff --staged --quiet; then
-            git commit -m "Update Mastodon posted tracker and backlinks [skip ci]"
-            git pull --rebase
-            git push
-          fi
